v0.8.11: code cleanup, HeaderBar always visible, lazy chop real-time slices
Priority 1 bugs:
- getTailLengthSeconds() returns 5.0 (max ADSR release); was 0 — VST contract violation
- GrainEngine BPM calc: keep double arithmetic, cast at end (was discarding precision)
- Voice muteGroup default changed from 0 to 1 to match slice creation default
- AdsrEnvelope: document 0.00015f (~-76 dB) and 1.01f (asymptotic curve) magic numbers

Design cleanup:
- UIHelpers.h (new): computeDragValue(), computeZoomFactor(), UILayout::waveformVerticalScale
- HeaderBar: extract paintRow1/paintRow2; use UIHelpers::computeDragValue()
- SliceControlBar: use UIHelpers::computeDragValue() (eliminate duplication)
- ScrollZoomBar: use UIHelpers::computeZoomFactor()
- WaveformView: encapsulate sliceDrawMode; CacheKey struct; extract paint helpers; use UIHelpers
- ActionPanel/PluginEditor: use WaveformView::setSliceDrawMode/isSliceDrawModeActive()
- PluginEditor: scaleDirty bool replaces -1.0f sentinel
- VoicePool: add const getVoice() overload; document audio constants

Features:
- HeaderBar now always shows parameter controls (Row1+Row2) even with no sample loaded;
  waveform area already shows drop hint so HeaderBar was redundant
- Lazy chop: slices now appear in the UI immediately on each MIDI note press instead
  of waiting until lazy chop is stopped (uiSnapshotDirty was gated behind midiSelectsSlice)

+++ src/audio/GrainEngine.cpp
@@ -4,5 +4,5 @@ float GrainEngine::calcStretchBpm (int startSample, int endSample, float timeUni
 {
     double durationSec = (endSample - startSample) / sampleRate;
     float beats = timeUnitBars * 4.0f;  // 4 beats per bar
-    return durationSec > 0.0 ? (beats / (float) durationSec) * 60.0f : 120.0f;
+    return durationSec > 0.0 ? static_cast<float> ((double) beats / durationSec * 60.0) : 120.0f;
 }
